Fix CTA button
I was using a helper class secondary that was setting the color of text.
This was confusing because the css framework also has a secondary class
on some elements that sets the background color. Changed helper class to
has-text-secondary to make it clearer on the purpose of the custom
helper.